patch (command)!: refactor to make structs built by resultBuilder encodable
greatly simplifies and declutters code. unfortunately a minor workaround was required to encode an array with protocol types as protocols don’t automatically generate codable conformance.

This leads to the error “AppCommand.swift:14:9: Cannot automatically synthesize 'Encodable' because '[CommandOption]' does not conform to 'Encodable’” if a custom encode func is not included